chore: Update socket and defaults files
In this commit, the files 'index.ts' from the 'Defaults' folder, 'messages-recv.ts' from the 'Socket/messages-recv.ts' folder, and 'Socket.ts' from the 'Types/Socket.ts' folder were modified. These changes were made to add 'forceGroupsPrekeys'. This update aims to improve the overall functionality of the project, as it enhances the socket communication and default settings.

Please note that the 'git restore --staged <file>' command can be used to unstage any of the modified files if necessary.