fix: Improve handling of jid in messages-recv.ts
This commit fixes an issue where the jid attribute was not being handled correctly in certain cases in the messages-recv.ts file. It has been updated to use a default value of 'from' when the jid attribute is not present. This will ensure that the jid is always defined, preventing potential errors. Additionally, the jidNormalizedUser function has been used to normalize the jid value. This change improves the reliability and robustness of the jid handling in this file.

Changes:
- src/Socket/messages-recv.ts