fix: Resolve 'jid is undefined' error in contacts.update
In the 'messages-recv.ts' file, the 'jid' variable was being passed as undefined when creating a new contact object in the 'contacts.update' event. This change resolves that issue by providing a default empty string value when 'jid' is undefined.

The 'id' field in the contact object will now be constructed using other available attributes such as 'from' or 'hash' when 'jid' is not present. This ensures that a valid 'id' value is always provided, avoiding any further errors.